consolidate valve_host and valve_flood in valve_switch.
